// src/types/entities.ts
export interface FileFromHttp {
  name: string; 
  type: string; 
  content: string; 
}